// src/data/types.ts
import { CSSProperties } from 'react'

export interface ProductLine {
    description: string
    quantity: string
    rate: string
}

export interface Invoice {
    name: string
    companyAddress: string
    companyAddress2: string
    companyCountry: string

    billTo: string
    clientName: string
    clientAddress: string
    clientAddress2: string
    clientCountry: string

    invoiceTitleLabel: string
    invoiceTitle: string
    invoiceDateLabel: string
    invoiceDate: string
    invoiceDueDateLabel: string
    invoiceDueDate: string

    productLineDescription: string
    productLineQuantity: string
    productLineQuantityRate: string
    productLineQuantityAmount: string

    productLines: ProductLine[]

    subTotalLabel: string

    totalLabel: string
    currency: string

    notesLabel: string
    notes: string
    termLabel: string
    term: string
}

export interface CSSClasses {
    [key: string]: CSSProperties
}